requirements.txt
pygame~=2.0.1
paho-mqtt~=1.5.1
Shapely~=1.7.1